requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
